fix and refactor tableComponent
