--- src/Pages/CheckEmail/checkEmail.module.scss
@@ -1,56 +1,0 @@
-@import '../../styles/mixins';
-
-
-.container {
-  @include animation-appearance(0.4s);
-
-  width: 420px;
-  height: 408px;
-  margin: 0 auto;
-  padding-top: 37px;
-
-  @media (max-width: 767px) {
-    display: block;
-    width: 360px;
-    padding: 33px 15px 25px;
-  }
-
-  .content {
-    display: flex;
-    flex-direction: column;
-    gap: 0 0;
-    align-items: center;
-    justify-content: flex-start;
-
-    padding: 37px 36px 48px;
-
-    .typographyHead {
-      margin-bottom: 30px;
-      font-size: var(--font-size-xxl);
-    }
-
-    .imageContainer {
-      display: flex;
-      justify-content: center;
-      width: 100%;
-      margin-bottom: 19px;
-
-      svg {
-        width: 96px;
-        height: 96px;
-        object-fit: contain;
-      }
-    }
-
-    .notification {
-      margin-bottom: 40px;
-      color: var(--color-light-900);
-      text-align: center;
-      opacity: 0.5;
-    }
-
-    .backBtn {
-      text-decoration: none;
-    }
-  }
-}
