+++ src/app/router/router.tsx
@@ -2,14 +2,14 @@
 import { Suspense } from 'react'
 import { Navigate, Outlet, RouteObject, createBrowserRouter } from 'react-router-dom'
 
-//import { CardsPage } from '@/Pages/CardsPage/Cards.page'
-import ErrorPage from '@/Pages/ErrorPage/ErrorPage'
 import { path } from '@/app/router/path'
 import { Layout } from '@/components/Layout/Layout'
 import Loading from '@/components/ui/Loading/Loading'
 import { useMeQuery } from '@/features/auth/services/auth.service'
 import { CheckEmail } from '@/features/auth/ui/CheckEmail/CheckEmail'
 import { CreateNewPassword } from '@/features/auth/ui/CreateNewPassword/CreateNewPassword'
+//import { CardsPage } from '@/Pages/CardsPage/Cards.page'
+import ErrorPage from '@/features/auth/ui/ErrorPage/ErrorPage'
 import { ForgotPassword } from '@/features/auth/ui/ForgotPassword/ForgotPassword'
 import { SignInPage } from '@/features/auth/ui/SingInPage/SignIn.page'
 import SignUp from '@/features/auth/ui/SingUp/SignUp'
