--- src/components/TableComponent/mobile/TableHeadMobile/tableHeadMobile.module.scss
@@ -1,48 +1,0 @@
-@import 'src/styles/mixins';
-
-.title {
-  width: 100%;
-  margin-bottom: 10px;
-  padding-left: 5px;
-  text-align: center;
-}
-
-.box {
-  display: flex;
-  justify-content: center;
-  width: 100%;
-  margin-bottom: 30px;
-}
-
-.btn {
-  padding: 0;
-  border: 0;
-  outline: 0;
-}
-
-.boxItem {
-  @include focus-visible-default;
-
-  cursor: pointer;
-
-  display: flex;
-  flex-flow: row nowrap;
-  align-items: center;
-  justify-content: center;
-
-  width: 100%;
-  min-width: max-content;
-  padding: 10px;
-
-  border: 1px solid var(--color-dark-300);
-}
-
-.arrow {
-  width: 24px;
-  height: 24px;
-  transition: transform 0.5s;
-}
-
-.rotate {
-  transform: rotate(180deg);
-}
