--- src/components/TableComponent/tableComponent.module.scss
@@ -1,84 +1,0 @@
-@import 'src/styles/mixins';
-
-.tableRoot {
-  table-layout: fixed;
-}
-
-//.emptyTableHeadCell {
-.lastTableHeadCell {
-  //width: 15%;
-  width: 10%;
-}
-
-.nameSortBtn {
-  @include focus-visible-default;
-
-  display: flex;
-}
-
-// Для Decks
-.tableHeadCellDecks:nth-child(1) {
-  //width: 35%;
-  width: 45%;
-  position: relative;
-}
-
-.tableHeadCellDecks:nth-child(2) {
-  width: 12%;
-  position: relative;
-}
-
-// Для Cards
-.tableHeadCellCards:nth-child(1),
-.tableHeadCellCards:nth-child(2) {
-  width: 50%;
-  position: relative;
-}
-.tableHeadCellCards:nth-child(3) {
-  width: 15%;
-  position: relative;
-}
-
-.arrow {
-  width: 24px;
-  height: 24px;
-  transition: transform 0.5s;
-}
-
-.rotate {
-  transform: rotate(180deg);
-}
-
-.empty {
-  padding: 60px 0;
-  text-align: center;
-}
-.answer {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-.boxEye {
-  display: flex;
-  justify-content: center;
-  width: 8%;
-  height: 100%;
-  transform: translate(0, 0);
-  position: absolute;
-  right: 0;
-  align-items: center;
-
-  &:focus-visible {
-    @include focus-visible-default;
-  }
-
-  svg g {
-    transition: fill 0.3s ease-in-out;
-  }
-
-  &:hover {
-    svg g {
-      fill: var(--color-accent-700);
-    }
-  }
-}
