+++ src/components/TableComponent/ui/TableComponent.module.scss
@@ -1,0 +1,85 @@
+@import '../../../styles/mixins';
+
+.tableRoot {
+  table-layout: fixed;
+}
+
+.lastTableHeadCell {
+  width: 10%;
+}
+
+.nameSortBtn {
+  @include focus-visible-default;
+  display: flex;
+}
+
+// Стили для Decks
+.tableHeadCellDecks {
+  &:nth-child(1) {
+    width: 45%;
+    position: relative;
+  }
+  &:nth-child(2) {
+    width: 12%;
+    position: relative;
+  }
+}
+
+// Стили для Cards
+.tableHeadCellCards {
+  &:nth-child(1),
+  &:nth-child(2) {
+    width: 50%;
+    position: relative;
+  }
+  &:nth-child(3) {
+    width: 15%;
+    position: relative;
+  }
+}
+
+.arrow {
+  width: 24px;
+  height: 24px;
+  transition: transform 0.5s;
+}
+
+.rotate {
+  transform: rotate(180deg);
+}
+
+.empty {
+  padding: 60px 0;
+  text-align: center;
+}
+
+.answer {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.boxEye {
+  display: flex;
+  justify-content: center;
+  width: 8%;
+  height: 100%;
+  transform: translate(0, 0);
+  position: absolute;
+  right: 0;
+  align-items: center;
+
+  &:focus-visible {
+    @include focus-visible-default;
+  }
+
+  svg g {
+    transition: fill 0.3s ease-in-out;
+  }
+
+  &:hover {
+    svg g {
+      fill: var(--color-accent-700);
+    }
+  }
+}
