+++ src/components/TableComponent/ui/mobile/TableCardMobile/TableCardMobile.module.scss
@@ -1,0 +1,173 @@
+@import '../../../../../styles/mixins';
+
+
+.root {
+  width: 100%;
+  max-width: 500px;
+  margin-bottom: 30px;
+
+  td {
+    text-align: right;
+  }
+}
+
+
+.imgWrapper {
+  @include focus-visible-default;
+
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  align-items: flex-end;
+  justify-content: start;
+
+  text-decoration: none;
+}
+
+.wrapperCoverImg {
+  max-width: 118px;
+  height: 48px;
+
+  .coverImg {
+    max-width: 118px;
+    height: 48px;
+
+    object-fit: cover;
+    object-position: center;
+
+    transition: transform 0.2s;
+
+    &:hover {
+      transform: scale(1.6);
+      height: auto;
+      object-fit: cover;
+    }
+  }
+}
+
+.grade {
+  display: flex;
+  flex-flow: row nowrap;
+  gap: 10px;
+  justify-content: flex-end;
+  align-items: center;
+
+  width: 100%;
+  height: 100%;
+
+  .star {
+    width: 16px;
+    height: 16px;
+    color: var(--color-warning-300);
+  }
+}
+
+.iconBtns {
+  position: relative;
+
+  display: flex;
+  flex-flow: row nowrap;
+  justify-content: center;
+
+  width: 100%;
+  gap: 40px;
+  padding: 20px;
+  border-left: 1px solid var(--color-dark-300);
+  border-right: 1px solid var(--color-dark-300);
+  border-bottom: 1px solid var(--color-dark-300);
+
+  .btn {
+    position: relative;
+
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    width: 16px;
+    padding: 0;
+
+    background-color: transparent;
+    border: none;
+    outline: none;
+    box-shadow: none;
+
+    &::before {
+      content: '';
+
+      position: absolute;
+      z-index: 0;
+      top: 50%;
+      left: 50%;
+      transform: translate(-50%, -50%) scale(0);
+
+      width: 28px;
+      height: 28px;
+
+      opacity: 0.5;
+      background-color: transparent;
+      border-radius: 50%;
+
+      transition: 0.5s;
+    }
+
+    &:focus-visible {
+      outline: none;
+    }
+
+    &:hover::before,
+    &:active::before,
+    &:focus-visible::before {
+      transform: translate(-50%, -50%) scale(1);
+      background-color: var(--color-dark-100);
+      outline: none;
+    }
+
+    &:focus-visible::before {
+      outline: 2px solid #71ff7a;
+    }
+
+    &:active::before {
+      background-color: var(--color-dark-300);
+    }
+
+    &:disabled::before {
+      background-color: transparent;
+    }
+  }
+}
+
+.img{
+  cursor: pointer;
+
+  position: absolute;
+
+  width: 16px;
+  height: 16px;
+
+  color: var(--color-light-100);
+
+  &:disabled {
+    background-color: transparent;
+  }
+}
+
+.disabled {
+  cursor: default;
+  opacity: 0.5;
+}
+.blur {
+  text-shadow: 0 0 5px var(--color-light-100);
+  opacity: 0.5;
+  transition: 0.1s all;
+}
+
+.blur p {
+  color: transparent;
+}
+
+.blur img {
+  filter: blur(4px);
+}
+.sell {
+  user-select: none;
+}
