+++ src/components/TableComponent/ui/singleRowCard/SingleRowCard.module.scss
@@ -1,0 +1,177 @@
+@import '../../../../styles/mixins';
+
+.imgWrapper {
+  @include focus-visible-default;
+
+  display: flex;
+  gap: 10px;
+  align-items: center;
+  justify-content: start;
+
+  text-decoration: none;
+
+  @media (max-width: 1200px) {
+    flex-direction: column;
+    justify-content: center;
+    align-items: start;
+  }
+}
+
+.wrapperCoverImg {
+  width: 118px;
+  height: 48px;
+  cursor: pointer;
+
+  .coverImg {
+    width: 118px;
+    height: 48px;
+    object-fit: cover;
+    object-position: center;
+    transition: transform 0.2s;
+
+    &:not(.blur) {
+      &:hover {
+        position: relative;
+        z-index: 2;
+        transform: scale(1.6);
+        height: auto;
+        object-fit: cover;
+
+        ~ .blur {
+          transform: none;
+          filter: blur(3px);
+        }
+      }
+    }
+  }
+
+  .withImg:not(.blur):hover {
+    transform: none;
+    filter: none;
+    height: 48px;
+  }
+}
+
+.grade {
+  display: flex;
+  flex-flow: row nowrap;
+  gap: 10px;
+  align-items: center;
+
+  width: 100%;
+  height: 100%;
+
+  .star {
+    width: 16px;
+    height: 16px;
+    color: var(--color-warning-300);
+  }
+}
+
+.iconBtns {
+  position: relative;
+
+  display: flex;
+  flex-flow: row nowrap;
+  gap: 10px;
+  justify-content: center;
+
+  .btn {
+    position: relative;
+
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    width: 16px;
+    padding: 0;
+
+    background-color: transparent;
+    border: none;
+    outline: none;
+    box-shadow: none;
+
+    &::before {
+      content: '';
+
+      position: absolute;
+      z-index: 0;
+      top: 50%;
+      left: 50%;
+      transform: translate(-50%, -50%) scale(0);
+
+      width: 28px;
+      height: 28px;
+
+      opacity: 0.5;
+      background-color: transparent;
+      border-radius: 50%;
+
+      transition: 0.5s;
+    }
+
+    &:focus-visible {
+      outline: none;
+    }
+
+    &:hover::before,
+    &:active::before,
+    &:focus-visible::before {
+      transform: translate(-50%, -50%) scale(1);
+      background-color: var(--color-dark-100);
+      outline: none;
+    }
+
+    &:focus-visible::before {
+      outline: 2px solid #71ff7a;
+    }
+
+    &:active::before {
+      background-color: var(--color-dark-300);
+    }
+
+    &:disabled::before {
+      background-color: transparent;
+    }
+  }
+}
+
+.Edit2Outline,
+.TrashOutline {
+  cursor: pointer;
+
+  position: absolute;
+
+  width: 16px;
+  height: 16px;
+
+  color: var(--color-light-100);
+
+  &:disabled {
+    background-color: transparent;
+  }
+}
+
+.blur {
+  cursor: pointer;
+  text-shadow: 0 0 5px var(--color-light-100);
+  opacity: 0.5;
+  transition: 0.1s all;
+}
+
+.blur p {
+  color: transparent;
+}
+
+.blur img {
+  filter: blur(4px);
+}
+
+.unBlur {
+  text-shadow: none;
+  transition: 0.1s all;
+}
+
+.sell {
+  user-select: none;
+}
