+++ src/components/TableComponent/ui/singleRowDeck/SingleRowDeck.module.scss
@@ -1,0 +1,57 @@
+@import '../../../../styles/mixins';
+
+.deckRowCell {
+  position: relative;
+  .privacyStep {
+    position: absolute;
+    right: 0;
+    transform: translateY(-50%);
+    top: 50%;
+    width: 20px;
+    opacity: 0.6;
+    display: flex;
+    align-items: center;
+    .privacyIcon {
+      width: 100%;
+      color: var(--color-accent-100);
+    }
+  }
+}
+
+.imgWrapper {
+  @include focus-visible-default;
+
+  display: flex;
+  gap: 10px;
+  align-items: center;
+  justify-content: start;
+
+  text-decoration: none;
+  .isMine {
+    color: var(--color-accent-500);
+  }
+}
+
+.wrapperCoverImg {
+  width: 118px;
+  height: 48px;
+  position: relative;
+
+  .coverImg {
+    width: 100%;
+    height: 100%;
+
+    object-fit: cover;
+    object-position: center;
+
+    transition: transform 0.2s;
+
+    &.withImg:hover {
+      transform: scale(1.6);
+      height: auto;
+      object-fit: cover;
+      position: absolute;
+      z-index: 1;
+    }
+  }
+}
