+++ src/components/TableComponent/ui/singleRowDeck/btns/RowDeckBtns.module.scss
