+++ src/features/decks/ui/DecksPage.module.scss
@@ -84,18 +84,6 @@
 
   width: 100%;
   max-width: 250px;
-
-  //.slider {
-  //  max-width: 150px;
-  //
-  //  @media (width <= 767px) {
-  //    max-width: 100%;
-  //  }
-  //}
-  //
-  //@media (width <= 767px) {
-  //  max-width: 100%;
-  //}
 }
 
 .clearFilterBox {
